Move encoding definition to encoding.py
